+++ packages/coding-agent/test/extensions-runner-b4-hooks.test.ts
@@ -66,6 +66,7 @@ describe("ExtensionRunner B4 hooks", () => {
 		setActiveTools: () => {},
 		setDeferredOverrides: () => {},
 		setToolNamespaces: () => {},
+		hydrateMessageDeliveredTools: () => {},
 		refreshTools: () => {},
 		getCommands: () => [],
 		setModel: async () => false,