+++ packages/coding-agent/test/extensions-runner-deferred-commands.test.ts
@@ -56,6 +56,7 @@ describe("ExtensionRunner deferred extension commands", () => {
 		setActiveTools: () => {},
 		setDeferredOverrides: () => {},
 		setToolNamespaces: () => {},
+		hydrateMessageDeliveredTools: () => {},
 		refreshTools: () => {},
 		getCommands: () => [],
 		setModel: async () => false,